fix(frontend): refactor layout
