+++ apps/frontend/src/components/annotationRightCorner/index.module.scss
@@ -4,8 +4,6 @@
   @extend .flexDirectionRow;
   justify-content: space-between;
   align-items: center;
-  width: calc(100vw - 560px);
-  width: 100%;
 }
 .left {
   width: 300px;
