+++ apps/frontend/src/pages/homepage/index.module.scss
@@ -1,24 +1,13 @@
 @import '../../utils/common/common.module';
-.outerFrame {
-  @extend .displayFlex;
-  @extend .flexDirectionColumn;
-  justify-content: flex-start;
-  align-items: center;
-  width: 100vw;
-  width: 100%;
-  height: 100vh;
+.main {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
 }
 
 .content {
-  @extend .displayFlex;
-  @extend .flexDirectionRow;
-  justify-content: center;
-  align-items: center;
-  height: calc(100vh - 56px);
-  height: calc(100vh + 56px);
-  width: 100%;
+  flex-grow: 1;
   background-color: #f4f5f7;
-  //background: #F4F5F7;
 }
 :global {
   .ant-modal-footer {
