--- apps/frontend/src/pages/login/index.less
@@ -1,20 +1,0 @@
-.login-page {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: #f0f2f5;
-
-  &-form {
-    width: 300px;
-    padding: 50px 40px 10px;
-    border-radius: 10px;
-
-    h2 {
-      text-align: center;
-    }
-
-    &_button {
-      width: 100%;
-    }
-  }
-}
