+++ apps/frontend/src/pages/login/index.module.scss
@@ -1,0 +1,7 @@
+@import '../../utils/common/common.module';
+.loginWrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
