+++ apps/frontend/src/pages/login/index.tsx
@@ -1,54 +1,15 @@
-import type { FC } from 'react';
-import { Button, Checkbox, Form, Input } from 'antd';
-import './index.less';
-import { useNavigate, useLocation } from 'react-router-dom';
-import { useDispatch } from 'react-redux';
+import React from 'react';
 
-import type { LoginParams } from '../../types/user/login';
-import { loginAsync } from '../../stores/user.store';
-import { formatSearch } from '../../utils/formatSearch';
-
-const initialValues: LoginParams = {
-  username: 'guest',
-  password: 'guest',
-  // remember: true
-};
-
-const LoginForm: FC = () => {
-  const navigate = useNavigate();
-  const location = useLocation();
-  const dispatch = useDispatch();
-
-  const onFinished = async (form: LoginParams) => {
-    const res = dispatch(await loginAsync(form));
-    if (!!res) {
-      const search = formatSearch(location.search);
-      const from = search.from || { pathname: '/' };
-      navigate(from);
-    }
-  };
-
+import currentStyles from './index.module.scss';
+import Login from '../../components/login';
+import Constants from '../../constants';
+import LogoTitle from '../../components/logoTitle';
+const LoginPage = () => {
   return (
-    <div className="login-page">
-      <Form<LoginParams> onFinish={onFinished} className="login-page-form" initialValues={initialValues}>
-        <h2>Label Bee Platform</h2>
-        <Form.Item name="username" rules={[{ required: true, message: '请输入用户名！' }]}>
-          <Input placeholder="用户名" />
-        </Form.Item>
-        <Form.Item name="password" rules={[{ required: true, message: '请输入密码！' }]}>
-          <Input type="password" placeholder="密码" />
-        </Form.Item>
-        <Form.Item name="remember" valuePropName="checked">
-          <Checkbox>记住用户</Checkbox>
-        </Form.Item>
-        <Form.Item>
-          <Button htmlType="submit" type="primary" className="login-page-form_button">
-            登录
-          </Button>
-        </Form.Item>
-      </Form>
+    <div className={currentStyles.loginWrapper}>
+      <LogoTitle />
+      <Login turnToSignUp={Constants.urlToRegister} turnToTaskList={Constants.urlToTasks} />
     </div>
   );
 };
-
-export default LoginForm;
+export default LoginPage;
