--- apps/frontend/src/pages/login1/index.module.scss
@@ -1,5 +1,0 @@
-@import '../../utils/common/common.module';
-.outerFrame {
-  @extend .displayFlex;
-  @extend .flexDirectionColumn;
-}
