+++ apps/frontend/src/routes/index.tsx
@@ -3,14 +3,14 @@ import type { RouteObject } from 'react-router';
 import { useRoutes } from 'react-router-dom';
 
 import AnnotationConfig from '../pages/annotationConfig';
-import Login1 from '../pages/login1/index';
+import Login1 from '../pages/login/index';
 import SignUp from '../pages/signUp';
 import Homepage from '../pages/homepage';
 import TaskList from '../pages/taskList';
 import CreateTask from '../pages/createTask';
 import InputInfoConfig from '../pages/inputInfoConfig';
 import InputData from '../pages/inputData';
-import TaskAnnotation from '../pages/annotation2';
+import TaskAnnotation from '../pages/annotation';
 import Samples from '../pages/samples';
 import RootGuard from '../pages/guards/rootGuard';
 import TaskSamplesFinished from '../pages/sampleFinished';
