+++ apps/frontend/src/routes/privateRoute.tsx
