--- apps/frontend/src/styles/antd.reset.less
@@ -1,39 +1,0 @@
-// .ant-list-item {
-//   padding: 12px 12px !important;
-// }
-
-// .ant-table-cell .ant-btn.ant-btn-link {
-//   padding: 0 4px !important;
-// }
-
-// .ant-btn + .ant-btn {
-//   margin-left: 10px;
-//   border-radius: 4px;
-// }
-
-// .ant-popover-inner-content {
-//   border-radius: 4px;
-// }
-
-// .ant-popover-inner {
-//   border-radius: 4px;
-// }
-// .ant-tabs-tab + .ant-tabs-tab {
-//   margin: 0 0 0 3px;
-// }
-
-// .ant-dropdown-menu {
-//   border-radius: 4px;
-//   padding: 12px 0px 0px;
-//   overflow: hidden;
-// }
-// .ant-dropdown-menu-item {
-//   margin-bottom: 12px;
-// }
-
-// .ant-modal-content {
-//   border-radius: 8px;
-// }
-// .ant-modal-header {
-//   border-radius: 8px 8px 0 0;
-// }
