+++ apps/frontend/src/styles/globalVariables.css
@@ -1,0 +1,3 @@
+:root {
+  --primary-color: #1b67ff;
+}
