+++ apps/frontend/src/styles/index.less
@@ -1,18 +1,1 @@
 @import url('./main.less');
-
-@import url('./antd.reset.less');
-
-.App {
-  display: -moz-flex;
-  display: flex;
-  display: -ms-flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
-//:global{
-//  .ant-modal-footer{
-//    padding-top:24px;
-//    padding-bottom:24px;
-//  }
-//}
