+++ apps/frontend/vite.config.ts
@@ -50,8 +50,9 @@ export default defineConfig({
       less: {
         javascriptEnabled: true, //注意，这一句是在less对象中，写在外边不起作用
         modifyVars: {
-          //在这里进行主题的修改，参考官方配置属性
-          // '@primary-color': '#1DA57A'
+          '@primary-color': '#1B67FF',
+          '@primary-color-hover': '#1B67FF',
+          '@primary-color-active': '#1B67FF',
         },
       },
     },
